FEAT: Column name, definition, nullable 추가 RENAME: 폴더 변경 REMOVE: 불필요한 폴더 삭제
