+++ src/main/java/com/estsoft/astronautbe/domain/dto/TrendResponseDTO.java
@@ -1,0 +1,4 @@
+package com.estsoft.astronautbe.domain.dto;
+
+public class TrendResponseDTO {
+}
