--- src/main/java/com/estsoft/astronautbe/domain/keyword/TrendResponseDTO.java
@@ -1,4 +1,0 @@
-package com.estsoft.astronautbe.domain.keyword;
-
-public class TrendResponseDTO {
-}
